refactor: move & split api calls out from util (#73)
* refactor: move apis out from util

* refactor: split & apply getting api url & api calls
