+++ src/apis/api.ts
@@ -1,0 +1,5 @@
+import axios from 'axios';
+
+export const apiClient = axios.create({
+  baseURL: import.meta.env.VITE_API_URL,
+});
