+++ src/apis/get-shmup-record-preview-list.ts
@@ -1,50 +1,8 @@
-import axios from 'axios';
-
-import {
-  GetShmupRecordArticleResponse,
-  GetShmupRecordPreviewListResponse,
-  ShmupRecord,
-  ShmupRecordPreview,
-} from '^/types';
+import { GetShmupRecordPreviewListResponse, ShmupRecordPreview } from '^/types';
 import { convertDateToString } from '^/utils/date-to-string';
+import { getAPIURL } from '^/utils/get-api-url';
 
-export const apiClient = axios.create({
-  baseURL: import.meta.env.VITE_API_URL,
-});
-
-export function getAPIURL(...paths: string[]) {
-  return paths.join('/');
-}
-
-export async function getShmupRecordArticle(
-  typeId: string,
-  recordId: string,
-  onLoad: (newIsLoading: boolean) => void,
-  onError: (newIsError: boolean) => void,
-  onComplete: (newShmupRecordArticle: ShmupRecord) => void
-) {
-  onLoad(true);
-  onError(false);
-
-  try {
-    const response = await apiClient.get<GetShmupRecordArticleResponse>(
-      getAPIURL('records', typeId, recordId)
-    );
-    onComplete({
-      ...response.data.data,
-      when: new Date(response.data.data.when),
-    });
-  } catch (error) {
-    /**
-     * 에러 출력이 필요하므로 no-console을 이 줄에 한해 해제함
-     */
-    // eslint-disable-next-line no-console
-    console.error(error);
-    onError(true);
-  }
-
-  onLoad(false);
-}
+import { apiClient } from './api';
 
 export async function getShmupRecordPreviewList(
   typeId: string,
