+++ src/utils/get-api-url.ts
@@ -1,0 +1,3 @@
+export function getAPIURL(...paths: string[]) {
+  return paths.join('/');
+}
